MainMenu and basic gui path
We now have a looping Gui system.
All guis inheritate from the abstract class Gui.
In this Way we can implement any number of gui with custom design and
behavior without having to modify main() or any other gui.

This will make expending the program way cleaner in the future.